+++ cpp/gradbench/evals/saddle.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "json.hpp"
+
+namespace saddle {
+// Will always be of length 2.
+typedef std::vector<double> Input;
+
+// Will always be of length 4.
+typedef std::vector<double> Output;
+
+
+}
