+++ tools/enzyme/Makefile
@@ -6,7 +6,7 @@ LDFLAGS=-fuse-ld=$(LLD) -O3 -fno-math-errno -flto -Wl,--load-pass-plugin=$(LLDEN
 ENZYME_LIB?=/gradbench/enzyme-build/Enzyme/
 LLDENZYME=$(ENZYME_LIB)/LLDEnzyme-19.so
 
-EXECUTABLES=run_hello run_gmm run_lstm run_ba run_ht run_kmeans run_particle
+EXECUTABLES=run_hello run_gmm run_lstm run_ba run_ht run_kmeans run_particle run_saddle
 
 all: $(EXECUTABLES)
