+++ tools/enzyme/evals.txt
@@ -5,3 +5,4 @@ ht
 kmeans
 lstm
 particle
+saddle
