+++ tools/enzyme/particle.py
@@ -1,0 +1,3 @@
+from gradbench.cpp_main import functions
+
+globals().update(functions(__file__, ["rr", "ff", "fr", "rf"]))
