+++ tools/enzyme/run_saddle.cpp
@@ -1,0 +1,10 @@
+#include "gradbench/evals/saddle.hpp"
+#include "gradbench/main.hpp"
+#include "enzyme.h"
+#include "solver.hpp"
+
+
+int main(int argc, char* argv[]) {
+  return generic_main(argc, argv, {
+    });
+}
